requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-search-documents==11.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.7.1
multidict==6.1.0
numpy==2.1.3
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
XlsxWriter==3.2.0
yarl==1.17.2
